correctly backpressure deflate; cleanup
fixes #14
