+++ lib/archiver/zip.js
@@ -7,35 +7,11 @@
  */
 
 var inherits = require('util').inherits;
-var zlib = require('zlib');
 
 var Archiver = require('./core');
 var headers = require('../headers/zip');
 var util = require('../util');
-
-function DeflateRawChecksum(options) {
-  zlib.DeflateRaw.call(this, options);
-
-  this.checksum = util.crc32.createCRC32();
-  this.digest = null;
-
-  this.rawSize = 0;
-
-  this.on('end', function() {
-    this.digest = this.checksum.digest();
-  });
-}
-
-inherits(DeflateRawChecksum, zlib.DeflateRaw);
-
-DeflateRawChecksum.prototype.write = function(chunk, cb) {
-  zlib.DeflateRaw.prototype.write.call(this, chunk, cb);
-
-  if (chunk) {
-    this.checksum.update(chunk);
-    this.rawSize += chunk.length;
-  }
-};
+var DeflateRawChecksum = require('../util/DeflateRawChecksum');
 
 var zipArchiver = module.exports = function(options) {
   Archiver.call(this);
@@ -123,35 +99,31 @@ zipArchiver.prototype._processFile = function(source, data, callback) {
 
   self.archiver.file = file;
 
-  var checksum = util.crc32.createCRC32();
-  var deflate;
 
   file.offset = self.archiver.pointer;
   self._push(headers.file.toBuffer(file));
 
+  var deflate;
+  
   if (file.store === false) {
     deflate = new DeflateRawChecksum(self.options.zlib);
 
     deflate.on('error', callback);
 
     deflate.on('data', function(chunk) {
-      file.compressedSize += chunk.length;
       self._push(chunk);
     });
 
     deflate.on('end', function() {
       file.crc32 = deflate.digest;
       file.uncompressedSize = deflate.rawSize;
+      file.compressedSize = deflate.compressedSize;
       onend();
     });
   }
 
-  function onend() {
-    if (file.store) {
-      file.crc32 = checksum.digest();
-      file.compressedSize = file.uncompressedSize;
-    }
 
+  function onend() {
     self._push(headers.descriptor.toBuffer(file));
 
     self.archiver.files.push(file);
@@ -164,7 +136,8 @@ zipArchiver.prototype._processFile = function(source, data, callback) {
   if (Buffer.isBuffer(source)) {
     if (file.store) {
       file.uncompressedSize += source.length;
-      checksum.update(source);
+      file.crc32 = util.crc32.createCRC32().update(source).digest();
+      file.compressedSize = file.uncompressedSize;
       self._push(source);
       onend();
     } else {
@@ -173,6 +146,8 @@ zipArchiver.prototype._processFile = function(source, data, callback) {
     }
   } else if (util.isStream(source)) {
     if (file.store) {
+      var checksum = util.crc32.createCRC32();
+
       source.on('error', callback);
 
       source.on('data', function(chunk) {
@@ -181,15 +156,17 @@ zipArchiver.prototype._processFile = function(source, data, callback) {
         self._push(chunk);
       });
 
-      source.on('end', onend);
+      source.on('end', function () {
+        file.compressedSize = file.uncompressedSize;
+        file.crc32 = checksum.digest();
+        onend()
+      });
+    } else {
+      source.pipe(deflate);
     }
 
     if (typeof source.resume === 'function') {
       source.resume();
-    }
-
-    if (file.store === false) {
-      source.pipe(deflate);
     }
   } else {
     callback(new Error('A valid Stream or Buffer instance is needed as input source'));
